+++ CHANGELOG.md
@@ -2,13 +2,32 @@
 
 - In the errorhandler class static methods were added.
 
+<br>
+
 # v0.1.3
 
 - _Set_ and _Get_ values have been added to the class
   properties.
 
+<br>
+
 # v0.1.4
 
-# New methods have been added:
+## New methods have been added:
 - add() // Add a list of numbers.
 - val() // Converts and you get the numeric value of the currency.
+
+<br>
+
+# v0.1.5
+
+## Now you can perform the basic Math operations 🧐
+<hr>
+
+### Update in the argument of the method [ add() ] move from a "List" => "number".
+- add(currency: number) // Add number
+
+### New methods have been added:
+- subtract()   // Substract number
+- multiply()  // Multiply number
+- divide()   // Divide number